Fix compiler warnings in libthread
main.c: sub '...' for 'va_list' to match declaration.
Remove some unused static functions.

kill.c: handle other enum cases by doing nothing.

channel.c: cast from volatile Channel when approriate.

threadimpl.h: specify int32_t instead of long.

thread.h: remove pragmas

libthread.json: turn on -Werror now that warnings are gone.

+++ sys/src/libthread/threadimpl.h
@@ -155,7 +155,7 @@ struct Ioproc
 	int tid;
 	Channel *c, *creply;
 	int inuse;
-	long (*op)(va_list*);
+	int32_t (*op)(va_list*);
 	va_list arg;
 	long ret;
 	char err[ERRMAX];
